+++ src/types.rs
@@ -18,6 +18,10 @@ pub(crate) type PublicKey = Arc<crate::ffi::PublicKey>;
 pub(crate) use lightning::ln::msgs::SocketAddress;
 #[cfg(feature = "uniffi")]
 pub(crate) type SocketAddress = Arc<crate::ffi::SocketAddress>;
+#[cfg(not(feature = "uniffi"))]
+pub(crate) use lightning::routing::gossip::NodeAlias;
+#[cfg(feature = "uniffi")]
+pub(crate) type NodeAlias = Arc<crate::ffi::NodeAlias>;
 use bitcoin::{OutPoint, ScriptBuf};
 use bitcoin_payment_instructions::amount::Amount as BPIAmount;
 use bitcoin_payment_instructions::dns_resolver::DNSHrnResolver;